requirements.txt
pyrebase
python-dotenv
Pillow